requirements.txt
aiohttp
discord
anyconf
html2text
bs4
requests
requests-toolbelt
pixivpy